Performed a little code refactoring with cleaning of comments
