+++ src/main/java/com/team404/kainosproject/controller/BandController.java
@@ -2,7 +2,6 @@ package com.team404.kainosproject.controller;
 
 import com.team404.kainosproject.model.Band;
 import com.team404.kainosproject.model.dto.BandDTO;
-import com.team404.kainosproject.model.dto.FinalBandDTO;
 import com.team404.kainosproject.service.BandService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -24,7 +23,7 @@ public class BandController {
     }
 
     @GetMapping("/bands-competencies")
-    public Iterable<FinalBandDTO> getBandDTOs(){
+    public Iterable<BandDTO> getBandDTOs(){
         return service.getAllBandsDTOs();
     }
