--- src/main/java/com/team404/kainosproject/model/dto/FinalBandDTO.java
@@ -1,33 +1,0 @@
-package com.team404.kainosproject.model.dto;
-
-import java.util.List;
-
-public class FinalBandDTO {
-
-    private String band;
-    private List<CompetencyDTO> competencies;
-
-    public FinalBandDTO(String band, List<CompetencyDTO> competencies) {
-        this.band = band;
-        this.competencies = competencies;
-    }
-
-    public FinalBandDTO() {
-    }
-
-    public String getBand() {
-        return band;
-    }
-
-    public void setBand(String band) {
-        this.band = band;
-    }
-
-    public List<CompetencyDTO> getCompetencies() {
-        return competencies;
-    }
-
-    public void setCompetencies(List<CompetencyDTO> competencies) {
-        this.competencies = competencies;
-    }
-}
