+++ src/main/java/com/team404/kainosproject/service/BandService.java
@@ -1,7 +1,6 @@
 package com.team404.kainosproject.service;
 
 import com.team404.kainosproject.model.Band;
-import com.team404.kainosproject.model.CompetencyIndicator;
 import com.team404.kainosproject.model.dto.*;
 import com.team404.kainosproject.repository.BandRepository;
 import org.slf4j.Logger;
@@ -9,10 +8,8 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.sql.SQLOutput;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 public class BandService {
@@ -43,20 +40,12 @@ public class BandService {
      *
      * @return Iterable of BandDTO objects
      */
-    public Iterable<FinalBandDTO> getAllBandsDTOs() {
+    public Iterable<BandDTO> getAllBandsDTOs() {
         Iterable<Band> bands = repository.findAll();
-        //List<BandDTO> bandDTOS = new ArrayList<>();
-
         List<String> competencies = new ArrayList<>();
         List<String> bandNames = new ArrayList<>();
 
         bands.forEach(e -> {
-            /*e.getCompetencyIndicators()
-                    .stream()
-                    .map(ce -> ce.getSubCompetency().getCompetency().getName())
-                    .distinct()
-                    .forEach(competencies::add);*/
-
             e.getCompetencyIndicators()
                     .stream()
                     .filter(ind -> !competencies.contains(ind.getSubCompetency().getCompetency().getName()))
@@ -67,22 +56,13 @@ public class BandService {
                     .map(ce -> ce.getBand().getName())
                     .distinct()
                     .forEach(bandNames::add);
-
-
-            /*List<CompetencyIndicatorDTO> ciDTOS = e.getCompetencyIndicators().stream()
-                    .map(ce -> new CompetencyIndicatorDTO(
-                            ce.getSubCompetency().getCompetency().getName(),
-                            ce.getDescription(),
-                            ce.getSubCompetency().getName())).collect(Collectors.toList());
-
-            bandDTOS.add(new BandDTO(e.getName(), ciDTOS));*/
         });
 
-        List<FinalBandDTO> finalBandDTOSList = new ArrayList<>();
+        List<BandDTO> bandDTOSList = new ArrayList<>();
 
         for(String bandName: bandNames){
-            FinalBandDTO finalBandDTO = new FinalBandDTO();
-            finalBandDTO.setBand(bandName);
+            BandDTO bandDTO = new BandDTO();
+            bandDTO.setBand(bandName);
             final List<CompetencyDTO> competencyDTOS = new ArrayList<>();
             for(String competency: competencies){
                 final List<IndicatorDTO> indicators = new ArrayList<>();
@@ -101,30 +81,10 @@ public class BandService {
                 }
             }
             LOG.info("[{}] BandDTO contains [{}] competencies", bandName, competencyDTOS.size());
-            finalBandDTO.setCompetencies(competencyDTOS);
-            finalBandDTOSList.add(finalBandDTO);
+            bandDTO.setCompetencies(competencyDTOS);
+            bandDTOSList.add(bandDTO);
         }
-        LOG.info("Created [{}] Band Data Transfer Objects from Band model", finalBandDTOSList.size());
-        return finalBandDTOSList;
+        LOG.info("Created [{}] Band Data Transfer Objects from Band model", bandDTOSList.size());
+        return bandDTOSList;
     }
-
 }
-
-/*for(BandDTO bandDTO: bandDTOS){
-            FinalBandDTO finalBand = new FinalBandDTO();
-            finalBand.setBand(bandDTO.getName());
-            List<CompetencyDTO> competencyDTOS = new ArrayList<>();
-            for(String competency : competencies){
-                List<IndicatorDTO> indicatorDTOSList = new ArrayList<>();
-
-                List<CompetencyIndicatorDTO> bandsByCompetencyIndicatorName =
-                        bandDTO.getCompetencyIndicators().stream().filter(e -> e.getCompetencyIndicatorName().equals(competency)).collect(Collectors.toList());
-
-                for(CompetencyIndicatorDTO cidto: bandsByCompetencyIndicatorName){
-                    indicatorDTOSList.add(new IndicatorDTO(cidto.getSubCompetencyName(), cidto.getDescription()));
-                }
-                competencyDTOS.add(new CompetencyDTO(competency, indicatorDTOSList));
-            }
-            finalBand.setCompetencies(competencyDTOS);
-            finalBandDTOSList.add(finalBand);
-        }*/
